Move FormCreateBranchAtRevision to CommandsDialogs and rename it to FormCreateBranch
