+++ GitUI/CommandsDialogs/FormCreateBranch.resx
