+++ GitUI/GitUI.csproj
@@ -572,11 +572,11 @@
     <Compile Include="CommandsDialogs\FormArchive.Designer.cs">
       <DependentUpon>FormArchive.cs</DependentUpon>
     </Compile>
-    <Compile Include="HelperDialogs\FormCreateBranchAtRevision.cs">
+    <Compile Include="CommandsDialogs\FormCreateBranch.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="HelperDialogs\FormCreateBranchAtRevision.Designer.cs">
-      <DependentUpon>FormCreateBranchAtRevision.cs</DependentUpon>
+    <Compile Include="CommandsDialogs\FormCreateBranch.Designer.cs">
+      <DependentUpon>FormCreateBranch.cs</DependentUpon>
     </Compile>
     <Compile Include="CommandsDialogs\BrowseDialog\FormChangeLog.cs">
       <SubType>Form</SubType>
@@ -1107,8 +1107,8 @@
       <DependentUpon>FormArchive.cs</DependentUpon>
       <SubType>Designer</SubType>
     </EmbeddedResource>
-    <EmbeddedResource Include="HelperDialogs\FormCreateBranchAtRevision.resx">
-      <DependentUpon>FormCreateBranchAtRevision.cs</DependentUpon>
+    <EmbeddedResource Include="CommandsDialogs\FormCreateBranch.resx">
+      <DependentUpon>FormCreateBranch.cs</DependentUpon>
       <SubType>Designer</SubType>
     </EmbeddedResource>
     <EmbeddedResource Include="CommandsDialogs\BrowseDialog\FormChangeLog.resx">
